Move compute_inner_ns from ASTFrame to W_Func (Part 1 of compute-ns plan)
W_Func.compute_inner_ns replaces ASTFrame.compute_ns, making it callable
from outside ASTFrame (e.g. from SPyVM.fast_call for __origin__ work).
ASTFrame now delegates to w_func.compute_inner_ns(args_w or []).
