+++ spy/tests/vm/test_function.py
@@ -6,7 +6,14 @@ from spy import ast
 from spy.fqn import FQN
 from spy.location import Loc
 from spy.vm.b import B
-from spy.vm.function import Color, FuncKind, FuncParam, LoweringStage, W_ASTFunc
+from spy.vm.function import (
+    Color,
+    FuncKind,
+    FuncParam,
+    LoweringStage,
+    W_ASTFunc,
+    W_BuiltinFunc,
+)
 from spy.vm.object import W_Type
 from spy.vm.vm import SPyVM
 from spy.vm.w import W_FuncType
@@ -131,6 +138,24 @@ class TestFunction:
         w_f2 = make_w_func("test::bar")
         with pytest.raises(AssertionError):
             w_f1.replace_with(w_f2)
+
+    def test_compute_inner_ns_no_type_args(self):
+        vm = SPyVM()
+        w_f = make_w_func("test::foo")
+        ns = w_f.compute_inner_ns([vm.wrap(42)])
+        # non-type arg is ignored: no qualifiers
+        assert ns == FQN("test::foo")
+
+    def test_compute_inner_ns_with_type_args(self):
+        w_f = make_w_func("test::add")
+        ns = w_f.compute_inner_ns([B.w_i32])
+        assert ns == FQN("test::add[i32]")
+
+    def test_compute_inner_ns_builtin(self):
+        w_functype = make_FuncType(B.w_i32, w_restype=B.w_i32, color="blue")
+        w_f = W_BuiltinFunc(w_functype, FQN("test::myfunc"), lambda vm, x: x)
+        ns = w_f.compute_inner_ns([B.w_i32])
+        assert ns == FQN("test::myfunc[i32]")
 
     def test_W_ASTFunc_repr(self):
         w_f1 = make_w_func("test::foo")
