+++ spy/vm/function.py
@@ -242,6 +242,36 @@ class W_Func(W_Object):
         FQN("builtins::type::__new__"),
     }
 
+    def compute_inner_ns(self, args_w: Sequence[W_Object]) -> FQN:
+        """
+        Try to generate a meaningful namespace for blue functions. The
+        idea is that if a blue func takes type parameters, we want to include
+        them in the qualifiers. E.g.:
+
+            @blue
+            def add(T):
+                def impl(x: T, y: T) -> T:
+                    return x + y
+                return impl
+
+            add(i32) # ==> add[i32]::impl
+            add(str) # ==> add[str]::impl
+
+        At the moment, the implementation is a bit ad-hoc and hackish, as it
+        considers ONLY type params as qualifiers, and ignores everything else.
+
+        Note that this is more about readability than correctness: in case of
+        blue params which are ignored, we might get clashing namespaces, but
+        this is still ok, because uniqueness of FQNs is guaranteed by
+        vm.get_unique_FQN().
+
+        This is fine as long as we don't support separate compilation. For sep
+        comp, we will probably need a deterministic and reproducible way to
+        compute unique FQNs out of a blue call.
+        """
+        quals = [w_arg.fqn for w_arg in args_w if isinstance(w_arg, W_Type)]
+        return self.fqn.with_qualifiers(quals)
+
     def spy_get_w_type(self, vm: "SPyVM") -> W_Type:
         return self.w_functype
